examples/sftp-automated-workflows-agentcore/modules/agentcore/requirements.txt
boto3>=1.34.0
strands-agents>=1.0.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
flask>=2.0.0
requests>=2.25.0
bedrock_agentcore
